refactor: improved internal types + more
- fix: convert `attach` and `detach` methods of `AnimatedArray` into method syntax
- fix: remove initializer from `payload` property of `AnimatedArray`
- chore: remove unused `is` functions
- chore: rename "src/types" modules to use .d.ts
- refactor: move `is.equ` into "src/animated/Controller.ts" and simplify it
- feat(ts): reworked interpolation types

--- TypeScript changes ---

Almost all of the changes in this commit do *not* impact the public API.

The `AnimatedInterpolation` class now has two type parameters:
  - In: The input values represented by an array
  - Out: The interpolated output value

The `Payload` type parameter of `AnimatedArray` must now be an array type.

The following types no longer implement the `SpringValue` protocol:
  - AnimatedValue
  - AnimatedValueArray
  - AnimatedInterpolation

The `AnimatedValue` class only deals with `number` types now, but the implementation still supports any type of value (note: only number values can be interpolated, though).

The `AnimatedValueArray` class now has a type parameter:
  - T: The animated tuple type of animated value types

Created the "src/types/common.ts" module, which holds several generic types.

The constructor and `update` method of `Controller` now allow `Partial<State>` to be passed in. The `next` function (which is passed to async scripts) also allows this now.

The `SpringValueFromInterpolation` interface (in "src/interpolate.ts") is now known as the `SpringInterpolator` interface. This is also the type of the `interpolate` function (exported by the public API).

Ensure the arrays passed into an `InterpolationConfig` object are readonly-compatible.

The `Out` type parameter of `InterpolationConfig` can now be an array of strings and/or numbers.

Renamed the following types in "src/types/animated":
  - AnimatableValue --> Animatable

Added the following types to "src/types/interpolation":
  - Interpolatable
  - Interpolator<In>
  - InterpolatorArgs<In, Out>
  - InterpolatorFn<In, Out>