feat: new "parent" prop + improved "attach" prop
The "attach" prop must be a function that returns a Controller object. We might let "parent" be a function someday, and we can deprecate this prop.

The "parent" prop must be a Controller object.

The "parent" prop is now cached in "this.props", instead of the "attach" prop.

Fixed some edge case with attaching.