fix: make "_start" method call "onEnd" immediately when 100% idle
This fixes the following edge case. An "update" function is passed to every async "to" prop. Calling this function is *not* guaranteed to start any animations. For example, if you pass {x: 0} when the animated "x" value is already 0, then there's no need to animate. In this case, the promise returned by the "update" function was never resolved, but now it is (thanks to this commit).