fix: change how existing "config" is sanitized
Previously, the existing config was thrown out entirely when the animation type was changed (eg: from a "duration" type to a "tension" type or a "frequency" type).

Now, we only clear the config props that would clash (based on the precedence of each config prop in the "SpringValue#advance" method).

Also, this means we never need to recreate the AnimationConfig object. :D