fix: useSprings with props function
This regression was introduced in the previous canary version.

It caused the props function to be called on every render.

The deps no longer have the "length" argument included by default, because we handle length changes in a separate useMemo call now.

If you want to update all springs when the length changes, you must include the length in your deps array explicitly.