fix: change how "from" prop behaves when "default" isn't true
If { from } is defined, then { reset: true } is implied. At the same time, if { to } is undefined, then { from } is used as the "to" value.

This makes { from } updates (w/o { to }) act as "set" method calls, except when default values are being updated.

Lastly, when { reset } is true or implied to be true, the Animated node now has its value set *after* any onRest handlers are called, rather than before.